+++ loading/parsing.py
@@ -56,15 +56,10 @@ def parseCommands(lines):
 
         for e in range(len(entries)):
             entry = entries[e]      # Iterate through each entry on the line
-
-            # Hopefully can remove by using str.split() instead of str.split(' ')
-            # # Remove artifacting
-            # while entry[0] == ' ':
-            #     entry = entry[1:]
-
             element = entry.split()[0]   # The element being called
             other = ' '.join(entry.split()[1:])  # Everything else, to be consumed bit by bit
             
+
             # Split at the argument separator and subsequently at spaces
             if other.count(argumentSequence) > 1:
                 raise ParseError('Only one instance of "{}" per entry allowed.'.format(argumentSequence))
@@ -89,82 +84,51 @@ def parseCommands(lines):
             else:
                 args = []
 
+
             # Split at the pin separator and subsequently at spaces
             # inputs and outputs should both be lists of strings
             if other.count(pinSeparatorSequence) > 1:
                 raise ParseError('Only one instance of "{}" per entry allowed.'.format(pinSeparatorSequence))
-            
             elif other.count(pinSeparatorSequence) == 1:
                 inputs, outputs = [o.split() for o in other.split(pinSeparatorSequence)]
-            
             else:
                 inputs = other.split()
                 outputs = []
 
-            # Detect bussing syntax errors
-            for i in inputs:
-                # print(i, i.count(bussingSequence[0]), i.count(bussingSequence[2]))
-                if i.count(bussingSequence[0]) > 1 or i.count(bussingSequence[2]) > 1:
-                    raise ParseError('Too many bussing operators: {}{}'.format(bussingSequence[0], bussingSequence[2]))
-                if i.count(bussingSequence[0]) != i.count(bussingSequence[2]):
-                    raise ParseError('Unmatched bussing operators: {}{}'.format(bussingSequence[0], bussingSequence[2]))
-            for o in outputs:
-                # print(o, o.count(bussingSequence[0]), o.count(bussingSequence[2]))
-                if o.count(bussingSequence[0]) > 1 or o.count(bussingSequence[2]) > 1:
-                    raise ParseError('Too many bussing operators: {}{}'.format(bussingSequence[0], bussingSequence[2]))
-                if o.count(bussingSequence[0]) != o.count(bussingSequence[2]):
-                    raise ParseError('Unmatched bussing operators: {}{}'.format(bussingSequence[0], bussingSequence[2]))
 
-            # Detect bussing (i.e. "data[0..25]")
-            for index, i in enumerate(inputs):
-                if bussingSequence[0] in i:
-                    name, bussing = i[0:-1].split(bussingSequence[0])
-                    # print(name, bussing)
-                    begin, end = bussing.split(bussingSequence[1])
-                    if len(begin) > len(end):
-                        length = len(begin)
-                    else:
-                        length = len(end)
-                    # print(begin, end)
-                    begin = int('0x' + begin, base=16)
-                    end = int('0x' + end, base=16)
-                    # print(begin, end)
+            for pins in [inputs, outputs]:
+                # Detect bussing syntax errors
+                for p in pins:
+                    # print(i, i.count(bussingSequence[0]), i.count(bussingSequence[2]))
+                    if p.count(bussingSequence[0]) > 1 or p.count(bussingSequence[2]) > 1:
+                        raise ParseError('Too many bussing operators: {}{}'.format(bussingSequence[0], bussingSequence[2]))
+                    if p.count(bussingSequence[0]) != p.count(bussingSequence[2]):
+                        raise ParseError('Unmatched bussing operators: {}{}'.format(bussingSequence[0], bussingSequence[2]))
+                
+                # Detect bussing (i.e. "data[0..25]")
+                for index, p in enumerate(pins):
+                    if bussingSequence[0] in p:
+                        name, bussing = p[0:-1].split(bussingSequence[0])
+                        # print(name, bussing)
+                        begin, end = bussing.split(bussingSequence[1])
+                        if len(begin) > len(end):
+                            length = len(begin)
+                        else:
+                            length = len(end)
+                        
+                        begin = int(begin)
+                        end = int(end)
+                        # print(begin, end, length)
 
-                    inputs[index] = name + hex(begin)[2:]
-                    for j in range(1, end - begin + 1):
-                        inputs.insert(index + j + 1, (name + hex(j)[2:]).zfill(length))
+                        pins[index] = name + str(begin).zfill(length)
+                        for j in range(1, end - begin + 1):
+                            # print(j)
+                            pins.insert(index + j + 1, (name + str(j + begin).zfill(length)))
 
-            for index, o in enumerate(outputs):
-                if bussingSequence[0] in o:
-                    name, bussing = o[0:-1].split(bussingSequence[0])
-                    # print(name, bussing)
-                    begin, end = bussing.split(bussingSequence[1])
-                    if len(begin) > len(end):
-                        length = len(begin)
-                    else:
-                        length = len(end)
-                    # print(begin, end)
-                    begin = int('0x' + begin, base=16)
-                    end = int('0x' + end, base=16)
-                    # print(begin, end)
-
-                    outputs[index] = name + hex(begin)[2:]
-                    for j in range(1, end - begin + 1):
-                        outputs.insert(index + j + 1, (name + hex(j)[2:]).zfill(length))
-
-
-            # Hopefully can remove by using str.split() instead of str.split(' ')
-            # # Remove artifacting
-            # while '' in inputs:
-            #     inputs.remove('')
-            # while '' in outputs:
-            #     outputs.remove('')
 
             if element == includeSequence:
                 if len(inputs) != 1 or len(outputs) != 1:
                     raise ParseError('{} must have file name and element name separated by {}'.format(includeSequence, pinSeparatorSequence))
-
-            
 
             commands.append(command(
                 l + lineNumberOffset,