+++ docs/releases.md
@@ -51,9 +51,17 @@ This document contains release notes for all versions from 0.1.0 on.
 * Improved PDB breakout by cleaning up what it prints when it returns.
 * Fixed labels not accessble from lgc code.
 
-### 0.8.0.pre
+### 0.8.0
 * Added support for strings in element arguments.
 * Added support for bussing in inputs and outputs.
 * Cleaned up debug printing.
 * Set up command line args.
 * Repalced `keys` with `Logical -k`.
+
+### 0.9.0.pre
+* Added Linux support
+* Added a log file
+* Removed verbose mode
+* Changed the key binding names
+// * Fixed loops not working
+* Upgraded to Python 3.9.9